requirements.txt
Django==2.2
fpdf==1.7.2
numpy==1.17.0
Pillow==6.1.0
pytz==2019.2
sqlparse==0.3.0
